oci-language-translation/requirements.txt
oci>=2.110.0
pyyaml
pandas